Adding worker report grace period
Because checking worker status and returning task
by a worker may be executed in parallel, there is
a possiblility that a worker already reported the task, but the master
first checks worker's status before processing result task.
To prevent this condition, task is considered as
dropped if task did not arrive within grace period
since a worker is not processing it.